refactor(runtime): split monolithic runtime.go into focused files
Split the 2,188-line pkg/runtime/runtime.go into four files by concern:

- runtime.go: core types, constructor, options, agent accessors, shared helpers
- loop.go: agent interaction loop (RunStream, Run), iteration control, tool
  registration, background agent handlers
- streaming.go: chat completion stream processing (handleStream),
  streamResult type, image stripping, token estimation
- tool_dispatch.go: tool execution pipeline (processToolCalls,
  executeWithApproval, permission checking, hook integration)

Also simplify the extracted code:
- Consolidate duplicate emitAgentWarnings/emitAgentWarningsWithSend into one
- Collapse identical thinking if/else branches into a single call
- Replace four background-agent handler wrappers with a generic adapter
- Extract recordAssistantMessage and compactIfNeeded from RunStream
- Extract pre/post tool hook logic from runTool into focused helpers

Pure refactoring with no behavioral changes.
