+++ pkg/runtime/runtime.go
@@ -1,41 +1,28 @@
 package runtime
 
 import (
-	"cmp"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log/slog"
 	"maps"
-	"slices"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
 	"github.com/modelcontextprotocol/go-sdk/mcp"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/docker/docker-agent/pkg/agent"
-	"github.com/docker/docker-agent/pkg/chat"
-	"github.com/docker/docker-agent/pkg/config/latest"
 	"github.com/docker/docker-agent/pkg/config/types"
 	"github.com/docker/docker-agent/pkg/hooks"
-	"github.com/docker/docker-agent/pkg/model/provider"
-	"github.com/docker/docker-agent/pkg/model/provider/options"
-	"github.com/docker/docker-agent/pkg/modelerrors"
 	"github.com/docker/docker-agent/pkg/modelsdev"
-	"github.com/docker/docker-agent/pkg/permissions"
 	"github.com/docker/docker-agent/pkg/rag"
 	ragtypes "github.com/docker/docker-agent/pkg/rag/types"
 	"github.com/docker/docker-agent/pkg/session"
 	"github.com/docker/docker-agent/pkg/sessiontitle"
 	"github.com/docker/docker-agent/pkg/team"
-	"github.com/docker/docker-agent/pkg/telemetry"
 	"github.com/docker/docker-agent/pkg/tools"
 	"github.com/docker/docker-agent/pkg/tools/builtin"
 	agenttool "github.com/docker/docker-agent/pkg/tools/builtin/agent"
@@ -222,18 +209,6 @@ type LocalRuntime struct {
 	onToolsChanged func(Event)
 
 	bgAgents *agenttool.Handler
-}
-
-type streamResult struct {
-	Calls             []tools.ToolCall
-	Content           string
-	ReasoningContent  string
-	ThinkingSignature string // Used with Anthropic's extended thinking feature
-	ThoughtSignature  []byte
-	Stopped           bool
-	ActualModel       string      // The actual model used (may differ from configured model with routing)
-	Usage             *chat.Usage // Token usage for this stream
-	RateLimit         *chat.RateLimit
 }
 
 type Opt func(*LocalRuntime)
@@ -869,7 +844,7 @@ func (r *LocalRuntime) EmitStartupInfo(ctx context.Context, sess *session.Sessio
 	}
 
 	// Emit agent warnings (if any) - these are quick
-	r.emitAgentWarningsWithSend(a, send)
+	r.emitAgentWarnings(a, func(e Event) { send(e) })
 
 	// Tool loading can be slow (MCP servers need to start)
 	// Emit progressive updates as each toolset loads
@@ -933,514 +908,6 @@ func (r *LocalRuntime) emitToolsProgressively(ctx context.Context, a *agent.Agen
 	send(ToolsetInfo(totalTools, false, r.CurrentAgentName()))
 }
 
-// registerDefaultTools registers the runtime-managed tool handlers.
-// The tool definitions themselves come from the agent's toolsets; this only
-// maps tool names to the runtime handler functions that implement them.
-func (r *LocalRuntime) registerDefaultTools() {
-	r.toolMap[builtin.ToolNameTransferTask] = r.handleTaskTransfer
-	r.toolMap[builtin.ToolNameHandoff] = r.handleHandoff
-	r.toolMap[builtin.ToolNameChangeModel] = r.handleChangeModel
-	r.toolMap[builtin.ToolNameRevertModel] = r.handleRevertModel
-	r.toolMap[agenttool.ToolNameRunBackgroundAgent] = r.handleRunBackgroundAgent
-	r.toolMap[agenttool.ToolNameListBackgroundAgents] = r.handleListBackgroundAgents
-	r.toolMap[agenttool.ToolNameViewBackgroundAgent] = r.handleViewBackgroundAgent
-	r.toolMap[agenttool.ToolNameStopBackgroundAgent] = r.handleStopBackgroundAgent
-}
-
-func (r *LocalRuntime) handleRunBackgroundAgent(ctx context.Context, sess *session.Session, tc tools.ToolCall, _ chan Event) (*tools.ToolCallResult, error) {
-	return r.bgAgents.HandleRun(ctx, sess, tc)
-}
-
-func (r *LocalRuntime) handleListBackgroundAgents(ctx context.Context, sess *session.Session, tc tools.ToolCall, _ chan Event) (*tools.ToolCallResult, error) {
-	return r.bgAgents.HandleList(ctx, sess, tc)
-}
-
-func (r *LocalRuntime) handleViewBackgroundAgent(ctx context.Context, sess *session.Session, tc tools.ToolCall, _ chan Event) (*tools.ToolCallResult, error) {
-	return r.bgAgents.HandleView(ctx, sess, tc)
-}
-
-func (r *LocalRuntime) handleStopBackgroundAgent(ctx context.Context, sess *session.Session, tc tools.ToolCall, _ chan Event) (*tools.ToolCallResult, error) {
-	return r.bgAgents.HandleStop(ctx, sess, tc)
-}
-
-func (r *LocalRuntime) finalizeEventChannel(ctx context.Context, sess *session.Session, events chan Event) {
-	// Clear the elicitation events channel before closing the events channel
-	// to prevent a send-on-closed-channel panic in elicitationHandler.
-	// Skip for background sessions (ToolsApproved=true) — they never set the
-	// channel, so clearing it would null out the parent session's channel.
-	if !sess.ToolsApproved {
-		r.clearElicitationEventsChannel()
-	}
-
-	defer close(events)
-
-	events <- StreamStopped(sess.ID, r.resolveSessionAgent(sess).Name())
-
-	r.executeOnUserInputHooks(ctx, sess.ID, "stream stopped")
-
-	telemetry.RecordSessionEnd(ctx)
-}
-
-// RunStream starts the agent's interaction loop and returns a channel of events
-func (r *LocalRuntime) RunStream(ctx context.Context, sess *session.Session) <-chan Event {
-	slog.Debug("Starting runtime stream", "agent", r.CurrentAgentName(), "session_id", sess.ID)
-	events := make(chan Event, 128)
-
-	go func() {
-		telemetry.RecordSessionStart(ctx, r.CurrentAgentName(), sess.ID)
-
-		ctx, sessionSpan := r.startSpan(ctx, "runtime.session", trace.WithAttributes(
-			attribute.String("agent", r.CurrentAgentName()),
-			attribute.String("session.id", sess.ID),
-		))
-		defer sessionSpan.End()
-
-		// Set the events channel for elicitation requests.
-		// Skip for background sessions (ToolsApproved=true): they have all tools
-		// pre-approved and will never trigger elicitation prompts. Setting the
-		// channel would overwrite the parent session's channel; clearing it at
-		// teardown would break any pending MCP auth flow in the parent.
-		if !sess.ToolsApproved {
-			r.setElicitationEventsChannel(events)
-		}
-
-		a := r.resolveSessionAgent(sess)
-
-		// Emit agent information for sidebar display
-		// Use getEffectiveModelID to account for active fallback cooldowns
-		events <- AgentInfo(a.Name(), r.getEffectiveModelID(a), a.Description(), a.WelcomeMessage())
-
-		// Emit team information
-		events <- TeamInfo(r.agentDetailsFromTeam(), a.Name())
-
-		// Initialize RAG and forward events
-		r.InitializeRAG(ctx, events)
-
-		r.emitAgentWarnings(a, events)
-		r.configureToolsetHandlers(a, events)
-
-		agentTools, err := r.getTools(ctx, a, sessionSpan, events)
-		if err != nil {
-			events <- Error(fmt.Sprintf("failed to get tools: %v", err))
-			return
-		}
-
-		events <- ToolsetInfo(len(agentTools), false, a.Name())
-
-		messages := sess.GetMessages(a)
-		if sess.SendUserMessage {
-			lastMsg := messages[len(messages)-1]
-			events <- UserMessage(lastMsg.Content, sess.ID, lastMsg.MultiContent, len(sess.Messages)-1)
-		}
-
-		events <- StreamStarted(sess.ID, a.Name())
-
-		defer r.finalizeEventChannel(ctx, sess, events)
-
-		r.registerDefaultTools()
-
-		iteration := 0
-		// Use a runtime copy of maxIterations so we don't modify the session's persistent config
-		runtimeMaxIterations := sess.MaxIterations
-
-		// toolModelOverride holds the per-toolset model from the most recent
-		// tool calls. It applies for one LLM turn, then resets.
-		var toolModelOverride string
-		var prevAgentName string
-
-		for {
-			a = r.resolveSessionAgent(sess)
-
-			// Clear per-tool model override on agent switch so it doesn't
-			// leak from one agent's toolset into another agent's turn.
-			if a.Name() != prevAgentName {
-				toolModelOverride = ""
-				prevAgentName = a.Name()
-			}
-
-			r.emitAgentWarnings(a, events)
-			r.configureToolsetHandlers(a, events)
-
-			agentTools, err := r.getTools(ctx, a, sessionSpan, events)
-			if err != nil {
-				events <- Error(fmt.Sprintf("failed to get tools: %v", err))
-				return
-			}
-
-			// Emit updated tool count. After a ToolListChanged MCP notification
-			// the cache is invalidated, so getTools above re-fetches from the
-			// server and may return a different count.
-			events <- ToolsetInfo(len(agentTools), false, a.Name())
-
-			// Check iteration limit
-			if runtimeMaxIterations > 0 && iteration >= runtimeMaxIterations {
-				slog.Debug(
-					"Maximum iterations reached",
-					"agent", a.Name(),
-					"iterations", iteration,
-					"max", runtimeMaxIterations,
-				)
-
-				events <- MaxIterationsReached(runtimeMaxIterations)
-
-				// Wait for user decision (resume / reject)
-				select {
-				case req := <-r.resumeChan:
-					if req.Type == ResumeTypeApprove {
-						slog.Debug("User chose to continue after max iterations", "agent", a.Name())
-						runtimeMaxIterations = iteration + 10
-					} else {
-						slog.Debug("User rejected continuation", "agent", a.Name())
-
-						assistantMessage := chat.Message{
-							Role: chat.MessageRoleAssistant,
-							Content: fmt.Sprintf(
-								"Execution stopped after reaching the configured max_iterations limit (%d).",
-								runtimeMaxIterations,
-							),
-							CreatedAt: time.Now().Format(time.RFC3339),
-						}
-
-						addAgentMessage(sess, a, &assistantMessage, events)
-						return
-					}
-
-				case <-ctx.Done():
-					slog.Debug(
-						"Context cancelled while waiting for resume confirmation",
-						"agent", a.Name(),
-						"session_id", sess.ID,
-					)
-					return
-				}
-			}
-
-			iteration++
-
-			// Exit immediately if the stream context has been cancelled (e.g., Ctrl+C)
-			if err := ctx.Err(); err != nil {
-				slog.Debug("Runtime stream context cancelled, stopping loop", "agent", a.Name(), "session_id", sess.ID)
-				return
-			}
-			slog.Debug("Starting conversation loop iteration", "agent", a.Name())
-
-			streamCtx, streamSpan := r.startSpan(ctx, "runtime.stream", trace.WithAttributes(
-				attribute.String("agent", a.Name()),
-				attribute.String("session.id", sess.ID),
-			))
-
-			model := a.Model()
-			defaultModelID := r.getEffectiveModelID(a)
-
-			// Per-tool model routing: use a cheaper model for this turn
-			// if the previous tool calls specified one, then reset.
-			if toolModelOverride != "" {
-				if overrideModel, err := r.resolveModelRef(ctx, toolModelOverride); err != nil {
-					slog.Warn("Failed to resolve per-tool model override; using agent default",
-						"model_override", toolModelOverride, "error", err)
-				} else {
-					slog.Info("Using per-tool model override for this turn",
-						"agent", a.Name(), "override", overrideModel.ID(), "primary", model.ID())
-					model = overrideModel
-				}
-				toolModelOverride = ""
-			}
-
-			// Apply thinking setting based on session state.
-			// When thinking is disabled: clone with thinking=false to clear any thinking config.
-			// When thinking is enabled: clone with thinking=true to ensure defaults are applied
-			// (this handles models with no thinking config, explicitly disabled thinking, or
-			// models that already have thinking configured).
-			if !sess.Thinking {
-				model = provider.CloneWithOptions(ctx, model, options.WithThinking(false))
-				slog.Debug("Cloned provider with thinking disabled", "agent", a.Name(), "model", model.ID())
-			} else {
-				// Always clone with thinking=true when session has thinking enabled.
-				// applyOverrides will apply provider defaults if ThinkingBudget is nil or disabled.
-				model = provider.CloneWithOptions(ctx, model, options.WithThinking(true))
-				slog.Debug("Cloned provider with thinking enabled", "agent", a.Name(), "model", model.ID())
-			}
-
-			modelID := model.ID()
-
-			// Notify sidebar when this turn uses a different model (per-tool override).
-			if modelID != defaultModelID {
-				events <- AgentInfo(a.Name(), modelID, a.Description(), a.WelcomeMessage())
-			}
-
-			slog.Debug("Using agent", "agent", a.Name(), "model", modelID)
-			slog.Debug("Getting model definition", "model_id", modelID)
-			m, err := r.modelsStore.GetModel(ctx, modelID)
-			if err != nil {
-				slog.Debug("Failed to get model definition", "error", err)
-			}
-
-			var contextLimit int64
-			if m != nil {
-				contextLimit = int64(m.Limit.Context)
-			}
-
-			if m != nil && r.sessionCompaction {
-				contextLength := sess.InputTokens + sess.OutputTokens
-				if contextLength > int64(float64(contextLimit)*0.9) {
-					r.Summarize(ctx, sess, "", events)
-				}
-			}
-
-			messages := sess.GetMessages(a)
-			slog.Debug("Retrieved messages for processing", "agent", a.Name(), "message_count", len(messages))
-
-			// Strip image content from messages if the model doesn't support image input.
-			// This prevents API errors when conversation history contains images (e.g. from
-			// tool results or user attachments) but the current model is text-only.
-			if m != nil && len(m.Modalities.Input) > 0 && !slices.Contains(m.Modalities.Input, "image") {
-				messages = stripImageContent(messages)
-			}
-
-			// Try primary model with fallback chain if configured
-			res, usedModel, err := r.tryModelWithFallback(streamCtx, a, model, messages, agentTools, sess, m, events)
-			if err != nil {
-				// Treat context cancellation as a graceful stop
-				if errors.Is(err, context.Canceled) {
-					slog.Debug("Model stream canceled by context", "agent", a.Name(), "session_id", sess.ID)
-					streamSpan.End()
-					return
-				}
-
-				// Auto-recovery: if the error is a context overflow and
-				// session compaction is enabled, compact the conversation
-				// and retry the request instead of surfacing raw errors.
-				if _, ok := errors.AsType[*modelerrors.ContextOverflowError](err); ok && r.sessionCompaction {
-					slog.Warn("Context window overflow detected, attempting auto-compaction",
-						"agent", a.Name(),
-						"session_id", sess.ID,
-						"input_tokens", sess.InputTokens,
-						"output_tokens", sess.OutputTokens,
-						"context_limit", contextLimit,
-					)
-					events <- Warning(
-						"The conversation has exceeded the model's context window. Automatically compacting the conversation history...",
-						a.Name(),
-					)
-					r.Summarize(ctx, sess, "", events)
-
-					// After compaction, loop back to retry with the
-					// compacted context. The next iteration will re-fetch
-					// messages from the (now compacted) session.
-					streamSpan.End()
-					continue
-				}
-
-				streamSpan.RecordError(err)
-				streamSpan.SetStatus(codes.Error, "error handling stream")
-				slog.Error("All models failed", "agent", a.Name(), "error", err)
-				// Track error in telemetry
-				telemetry.RecordError(ctx, err.Error())
-				events <- Error(modelerrors.FormatError(err))
-				streamSpan.End()
-				return
-			}
-
-			// Update sidebar model info to reflect what was actually used this turn.
-			// Fallback models are sticky (cooldown system persists them), so we only
-			// emit once. Per-tool model overrides are temporary (one turn), so we
-			// emit the override and then revert to the agent's default.
-			if usedModel != nil && usedModel.ID() != model.ID() {
-				slog.Info("Used fallback model", "agent", a.Name(), "primary", model.ID(), "used", usedModel.ID())
-				events <- AgentInfo(a.Name(), usedModel.ID(), a.Description(), a.WelcomeMessage())
-			} else if model.ID() != defaultModelID {
-				// Per-tool override was active: revert sidebar to the agent's default model.
-				events <- AgentInfo(a.Name(), defaultModelID, a.Description(), a.WelcomeMessage())
-			}
-			streamSpan.SetAttributes(
-				attribute.Int("tool.calls", len(res.Calls)),
-				attribute.Int("content.length", len(res.Content)),
-				attribute.Bool("stopped", res.Stopped),
-			)
-			streamSpan.End()
-			slog.Debug("Stream processed", "agent", a.Name(), "tool_calls", len(res.Calls), "content_length", len(res.Content), "stopped", res.Stopped)
-
-			// Add assistant message to conversation history, but skip empty assistant messages
-			// Providers reject assistant messages that have neither content nor tool calls.
-			var msgUsage *MessageUsage
-			if strings.TrimSpace(res.Content) != "" || len(res.Calls) > 0 {
-				// Build tool definitions for the tool calls
-				var toolDefs []tools.Tool
-				if len(res.Calls) > 0 {
-					toolMap := make(map[string]tools.Tool, len(agentTools))
-					for _, t := range agentTools {
-						toolMap[t.Name] = t
-					}
-					for _, call := range res.Calls {
-						if def, ok := toolMap[call.Function.Name]; ok {
-							toolDefs = append(toolDefs, def)
-						}
-					}
-				}
-
-				// Calculate per-message cost if usage and pricing info available
-				var messageCost float64
-				if res.Usage != nil && m != nil && m.Cost != nil {
-					messageCost = (float64(res.Usage.InputTokens)*m.Cost.Input +
-						float64(res.Usage.OutputTokens)*m.Cost.Output +
-						float64(res.Usage.CachedInputTokens)*m.Cost.CacheRead +
-						float64(res.Usage.CacheWriteTokens)*m.Cost.CacheWrite) / 1e6
-				}
-
-				// Determine the model name to store
-				messageModel := cmp.Or(res.ActualModel, modelID)
-
-				assistantMessage := chat.Message{
-					Role:              chat.MessageRoleAssistant,
-					Content:           res.Content,
-					ReasoningContent:  res.ReasoningContent,
-					ThinkingSignature: res.ThinkingSignature,
-					ThoughtSignature:  res.ThoughtSignature,
-					ToolCalls:         res.Calls,
-					ToolDefinitions:   toolDefs,
-					CreatedAt:         time.Now().Format(time.RFC3339),
-					Usage:             res.Usage,
-					Model:             messageModel,
-					Cost:              messageCost,
-				}
-
-				// Build per-message usage for the event
-				if res.Usage != nil {
-					msgUsage = &MessageUsage{
-						Usage: *res.Usage,
-						Cost:  messageCost,
-						Model: messageModel,
-					}
-					if res.RateLimit != nil {
-						msgUsage.RateLimit = *res.RateLimit
-					}
-				}
-
-				addAgentMessage(sess, a, &assistantMessage, events)
-				slog.Debug("Added assistant message to session", "agent", a.Name(), "total_messages", len(sess.GetAllMessages()))
-			} else {
-				slog.Debug("Skipping empty assistant message (no content and no tool calls)", "agent", a.Name())
-			}
-
-			usage := SessionUsage(sess, contextLimit)
-			usage.LastMessage = msgUsage
-			events <- NewTokenUsageEvent(sess.ID, a.Name(), usage)
-
-			// Record the message count before tool calls so we can
-			// measure how much content was added by tool results.
-			messageCountBeforeTools := len(sess.GetAllMessages())
-
-			r.processToolCalls(ctx, sess, res.Calls, agentTools, events)
-
-			// Record per-toolset model override for the next LLM turn.
-			toolModelOverride = resolveToolCallModelOverride(res.Calls, agentTools)
-
-			if res.Stopped {
-				slog.Debug("Conversation stopped", "agent", a.Name())
-				break
-			}
-
-			// Root-cause fix for stale token counts (issue #1750):
-			// After tool calls, sess.InputTokens still reflects the
-			// *previous* API response and doesn't account for the
-			// (potentially large) tool results just added. Estimate
-			// the additional tokens and compact proactively to prevent
-			// the oversized request from ever being sent.
-			if m != nil && r.sessionCompaction && contextLimit > 0 {
-				newMessages := sess.GetAllMessages()[messageCountBeforeTools:]
-				var addedTokens int64
-				for _, msg := range newMessages {
-					addedTokens += estimateMessageTokens(&msg.Message)
-				}
-
-				estimatedTotal := sess.InputTokens + sess.OutputTokens + addedTokens
-				if estimatedTotal > int64(float64(contextLimit)*0.9) {
-					slog.Info("Proactive compaction: tool results pushed estimated context past 90%% threshold",
-						"agent", a.Name(),
-						"input_tokens", sess.InputTokens,
-						"output_tokens", sess.OutputTokens,
-						"added_estimated_tokens", addedTokens,
-						"estimated_total", estimatedTotal,
-						"context_limit", contextLimit,
-					)
-					r.Summarize(ctx, sess, "", events)
-				}
-			}
-		}
-	}()
-
-	return events
-}
-
-// getTools executes tool retrieval with automatic OAuth handling
-func (r *LocalRuntime) getTools(ctx context.Context, a *agent.Agent, sessionSpan trace.Span, events chan Event) ([]tools.Tool, error) {
-	shouldEmitMCPInit := len(a.ToolSets()) > 0
-	if shouldEmitMCPInit {
-		events <- MCPInitStarted(a.Name())
-	}
-	defer func() {
-		if shouldEmitMCPInit {
-			events <- MCPInitFinished(a.Name())
-		}
-	}()
-
-	agentTools, err := a.Tools(ctx)
-	if err != nil {
-		slog.Error("Failed to get agent tools", "agent", a.Name(), "error", err)
-		sessionSpan.RecordError(err)
-		sessionSpan.SetStatus(codes.Error, "failed to get tools")
-		telemetry.RecordError(ctx, err.Error())
-		return nil, err
-	}
-
-	slog.Debug("Retrieved agent tools", "agent", a.Name(), "tool_count", len(agentTools))
-	return agentTools, nil
-}
-
-// configureToolsetHandlers sets up elicitation and OAuth handlers for all toolsets of an agent.
-func (r *LocalRuntime) configureToolsetHandlers(a *agent.Agent, events chan Event) {
-	for _, toolset := range a.ToolSets() {
-		tools.ConfigureHandlers(toolset,
-			r.elicitationHandler,
-			func() { events <- Authorization(tools.ElicitationActionAccept, a.Name()) },
-			r.managedOAuth,
-		)
-	}
-}
-
-// emitAgentWarningsWithSend emits agent warnings using the provided send function for context-aware sending.
-func (r *LocalRuntime) emitAgentWarningsWithSend(a *agent.Agent, send func(Event) bool) {
-	warnings := a.DrainWarnings()
-	if len(warnings) == 0 {
-		return
-	}
-
-	slog.Warn("Tool setup partially failed; continuing", "agent", a.Name(), "warnings", warnings)
-	send(Warning(formatToolWarning(a, warnings), a.Name()))
-}
-
-func (r *LocalRuntime) emitAgentWarnings(a *agent.Agent, events chan Event) {
-	warnings := a.DrainWarnings()
-	if len(warnings) == 0 {
-		return
-	}
-
-	slog.Warn("Tool setup partially failed; continuing", "agent", a.Name(), "warnings", warnings)
-
-	if events != nil {
-		events <- Warning(formatToolWarning(a, warnings), a.Name())
-	}
-}
-
-func formatToolWarning(a *agent.Agent, warnings []string) string {
-	var builder strings.Builder
-	fmt.Fprintf(&builder, "Some toolsets failed to initialize for agent '%s'.\n\nDetails:\n\n", a.Name())
-	for _, warning := range warnings {
-		fmt.Fprintf(&builder, "- %s\n", warning)
-	}
-	return strings.TrimSuffix(builder.String(), "\n")
-}
-
 func (r *LocalRuntime) Resume(_ context.Context, req ResumeRequest) {
 	slog.Debug("Resuming runtime", "agent", r.CurrentAgentName(), "type", req.Type, "reason", req.Reason)
 
@@ -1500,581 +967,7 @@ func (r *LocalRuntime) ResumeElicitation(ctx context.Context, action tools.Elici
 }
 
 // Run starts the agent's interaction loop
-func (r *LocalRuntime) Run(ctx context.Context, sess *session.Session) ([]session.Message, error) {
-	eventsChan := r.RunStream(ctx, sess)
 
-	for event := range eventsChan {
-		if errEvent, ok := event.(*ErrorEvent); ok {
-			return nil, fmt.Errorf("%s", errEvent.Error)
-		}
-	}
-
-	return sess.GetAllMessages(), nil
-}
-
-func (r *LocalRuntime) handleStream(ctx context.Context, stream chat.MessageStream, a *agent.Agent, agentTools []tools.Tool, sess *session.Session, m *modelsdev.Model, events chan Event) (streamResult, error) {
-	defer stream.Close()
-
-	var fullContent strings.Builder
-	var fullReasoningContent strings.Builder
-	var thinkingSignature string
-	var thoughtSignature []byte
-	var toolCalls []tools.ToolCall
-	var actualModel string
-	var messageUsage *chat.Usage
-	var messageRateLimit *chat.RateLimit
-
-	toolCallIndex := make(map[string]int)   // toolCallID -> index in toolCalls slice
-	emittedPartial := make(map[string]bool) // toolCallID -> whether we've emitted a partial event
-	toolDefMap := make(map[string]tools.Tool, len(agentTools))
-	for _, t := range agentTools {
-		toolDefMap[t.Name] = t
-	}
-
-	// recordUsage persists the final token counts and emits telemetry exactly
-	// once per stream, after we have the most accurate usage snapshot.
-	usageRecorded := false
-	recordUsage := func() {
-		if usageRecorded || messageUsage == nil {
-			return
-		}
-		usageRecorded = true
-
-		sess.InputTokens = messageUsage.InputTokens + messageUsage.CachedInputTokens + messageUsage.CacheWriteTokens
-		sess.OutputTokens = messageUsage.OutputTokens
-
-		modelName := "unknown"
-		if m != nil {
-			modelName = m.Name
-		}
-		telemetry.RecordTokenUsage(ctx, modelName, sess.InputTokens, sess.OutputTokens, sess.TotalCost())
-	}
-
-	for {
-		response, err := stream.Recv()
-		if errors.Is(err, io.EOF) {
-			break
-		}
-		if err != nil {
-			return streamResult{Stopped: true}, fmt.Errorf("error receiving from stream: %w", err)
-		}
-
-		if response.Usage != nil {
-			// Always keep the latest usage snapshot; some providers (e.g.
-			// Gemini) emit updated usage on every chunk with cumulative
-			// token counts, so the last value is the most accurate.
-			messageUsage = response.Usage
-		}
-
-		if response.RateLimit != nil {
-			messageRateLimit = response.RateLimit
-		}
-
-		if len(response.Choices) == 0 {
-			continue
-		}
-		choice := response.Choices[0]
-
-		if len(choice.Delta.ThoughtSignature) > 0 {
-			thoughtSignature = choice.Delta.ThoughtSignature
-		}
-
-		// Capture the actual model from the stream response (useful for model routing)
-		if actualModel == "" && response.Model != "" {
-			actualModel = response.Model
-		}
-
-		if choice.FinishReason == chat.FinishReasonStop || choice.FinishReason == chat.FinishReasonLength {
-			recordUsage()
-			return streamResult{
-				Calls:             toolCalls,
-				Content:           fullContent.String(),
-				ReasoningContent:  fullReasoningContent.String(),
-				ThinkingSignature: thinkingSignature,
-				ThoughtSignature:  thoughtSignature,
-				Stopped:           true,
-				ActualModel:       actualModel,
-				Usage:             messageUsage,
-				RateLimit:         messageRateLimit,
-			}, nil
-		}
-
-		// Handle tool calls
-		if len(choice.Delta.ToolCalls) > 0 {
-			// Process each tool call delta
-			for _, delta := range choice.Delta.ToolCalls {
-				idx, exists := toolCallIndex[delta.ID]
-				if !exists {
-					idx = len(toolCalls)
-					toolCallIndex[delta.ID] = idx
-					toolCalls = append(toolCalls, tools.ToolCall{
-						ID:   delta.ID,
-						Type: delta.Type,
-					})
-				}
-
-				tc := &toolCalls[idx]
-
-				// Track if we're learning the name for the first time
-				learningName := delta.Function.Name != "" && tc.Function.Name == ""
-
-				// Update fields from delta
-				if delta.Type != "" {
-					tc.Type = delta.Type
-				}
-				if delta.Function.Name != "" {
-					tc.Function.Name = delta.Function.Name
-				}
-				if delta.Function.Arguments != "" {
-					tc.Function.Arguments += delta.Function.Arguments
-				}
-
-				// Emit PartialToolCall once we have a name, and on subsequent argument deltas
-				if tc.Function.Name != "" && (learningName || delta.Function.Arguments != "") {
-					if !emittedPartial[delta.ID] || delta.Function.Arguments != "" {
-						events <- PartialToolCall(*tc, toolDefMap[tc.Function.Name], a.Name())
-						emittedPartial[delta.ID] = true
-					}
-				}
-			}
-			continue
-		}
-
-		if choice.Delta.ReasoningContent != "" {
-			events <- AgentChoiceReasoning(a.Name(), choice.Delta.ReasoningContent)
-			fullReasoningContent.WriteString(choice.Delta.ReasoningContent)
-		}
-
-		// Capture thinking signature for Anthropic extended thinking
-		if choice.Delta.ThinkingSignature != "" {
-			thinkingSignature = choice.Delta.ThinkingSignature
-		}
-
-		if choice.Delta.Content != "" {
-			events <- AgentChoice(a.Name(), choice.Delta.Content)
-			fullContent.WriteString(choice.Delta.Content)
-		}
-	}
-
-	recordUsage()
-
-	// If the stream completed without producing any content or tool calls, likely because of a token limit, stop to avoid breaking the request loop
-	// NOTE(krissetto): this can likely be removed once compaction works properly with all providers (aka dmr)
-	stoppedDueToNoOutput := fullContent.Len() == 0 && len(toolCalls) == 0
-	return streamResult{
-		Calls:             toolCalls,
-		Content:           fullContent.String(),
-		ReasoningContent:  fullReasoningContent.String(),
-		ThinkingSignature: thinkingSignature,
-		ThoughtSignature:  thoughtSignature,
-		Stopped:           stoppedDueToNoOutput,
-		ActualModel:       actualModel,
-		Usage:             messageUsage,
-		RateLimit:         messageRateLimit,
-	}, nil
-}
-
-// processToolCalls handles the execution of tool calls for an agent
-func (r *LocalRuntime) processToolCalls(ctx context.Context, sess *session.Session, calls []tools.ToolCall, agentTools []tools.Tool, events chan Event) {
-	a := r.resolveSessionAgent(sess)
-	slog.Debug("Processing tool calls", "agent", a.Name(), "call_count", len(calls))
-
-	// Build a map of agent tools for quick lookup
-	agentToolMap := make(map[string]tools.Tool, len(agentTools))
-	for _, t := range agentTools {
-		agentToolMap[t.Name] = t
-	}
-
-	for _, toolCall := range calls {
-		callCtx, callSpan := r.startSpan(ctx, "runtime.tool.call", trace.WithAttributes(
-			attribute.String("tool.name", toolCall.Function.Name),
-			attribute.String("tool.type", string(toolCall.Type)),
-			attribute.String("agent", a.Name()),
-			attribute.String("session.id", sess.ID),
-			attribute.String("tool.call_id", toolCall.ID),
-		))
-
-		slog.Debug("Processing tool call", "agent", a.Name(), "tool", toolCall.Function.Name, "session_id", sess.ID)
-
-		// Resolve the tool: it must be in the agent's tool set to be callable.
-		// After a handoff the model may hallucinate tools it saw in the
-		// conversation history from a previous agent; rejecting unknown
-		// tools with an error response lets it self-correct.
-		tool, available := agentToolMap[toolCall.Function.Name]
-		if !available {
-			slog.Warn("Tool call for unavailable tool", "agent", a.Name(), "tool", toolCall.Function.Name, "session_id", sess.ID)
-			errTool := tools.Tool{Name: toolCall.Function.Name}
-			r.addToolErrorResponse(ctx, sess, toolCall, errTool, events, a, fmt.Sprintf("Tool '%s' is not available. You can only use the tools provided to you.", toolCall.Function.Name))
-			callSpan.SetStatus(codes.Error, "tool not available")
-			callSpan.End()
-			continue
-		}
-
-		// Pick the handler: runtime-managed tools (transfer_task, handoff)
-		// have dedicated handlers; everything else goes through the toolset.
-		var runTool func()
-		if handler, exists := r.toolMap[toolCall.Function.Name]; exists {
-			runTool = func() { r.runAgentTool(callCtx, handler, sess, toolCall, tool, events, a) }
-		} else {
-			runTool = func() { r.runTool(callCtx, tool, toolCall, events, sess, a) }
-		}
-
-		// Execute tool with approval check
-		canceled := r.executeWithApproval(callCtx, sess, toolCall, tool, events, a, runTool)
-		if canceled {
-			callSpan.SetStatus(codes.Ok, "tool call canceled by user")
-			callSpan.End()
-			return
-		}
-
-		callSpan.SetStatus(codes.Ok, "tool call processed")
-		callSpan.End()
-	}
-}
-
-// executeWithApproval handles the tool approval flow and executes the tool.
-// Returns true if the operation was canceled and processing should stop.
-//
-// The approval flow considers (in order):
-//
-//  1. sess.ToolsApproved (--yolo flag) - auto-approve everything, takes precedence
-//  2. Session-level permissions (if configured) - pattern-based Allow/Ask/Deny rules
-//  3. Team-level permissions config - checked second
-//  4. Read-only hint - auto-approve
-//  5. Default: ask for user confirmation
-func (r *LocalRuntime) executeWithApproval(
-	ctx context.Context,
-	sess *session.Session,
-	toolCall tools.ToolCall,
-	tool tools.Tool,
-	events chan Event,
-	a *agent.Agent,
-	runTool func(),
-) (canceled bool) {
-	toolName := toolCall.Function.Name
-
-	// --yolo flag takes absolute precedence: auto-approve everything.
-	if sess.ToolsApproved {
-		slog.Debug("Tool auto-approved by --yolo flag", "tool", toolName, "session_id", sess.ID)
-		runTool()
-		return false
-	}
-
-	// Parse tool arguments once for permission matching
-	var toolArgs map[string]any
-	if toolCall.Function.Arguments != "" {
-		if err := json.Unmarshal([]byte(toolCall.Function.Arguments), &toolArgs); err != nil {
-			slog.Debug("Failed to parse tool arguments for permission check", "tool", toolName, "error", err)
-			// Continue with nil args - will only match tool name patterns
-		}
-	}
-
-	// Collect permission checkers in priority order (session first, then team)
-	checkers := r.permissionCheckers(sess)
-
-	for _, pc := range checkers {
-		switch pc.checker.CheckWithArgs(toolName, toolArgs) {
-		case permissions.Deny:
-			slog.Debug("Tool denied by permissions", "tool", toolName, "source", pc.source, "session_id", sess.ID)
-			r.addToolErrorResponse(ctx, sess, toolCall, tool, events, a, fmt.Sprintf("Tool '%s' is denied by %s.", toolName, pc.source))
-			return false
-		case permissions.Allow:
-			slog.Debug("Tool auto-approved by permissions", "tool", toolName, "source", pc.source, "session_id", sess.ID)
-			runTool()
-			return false
-		case permissions.ForceAsk:
-			slog.Debug("Tool requires confirmation (ask pattern)", "tool", toolName, "source", pc.source, "session_id", sess.ID)
-			return r.askUserForConfirmation(ctx, sess, toolCall, tool, events, a, runTool)
-		case permissions.Ask:
-			// No explicit match at this level; fall through to next checker
-		}
-	}
-
-	// No permission rule matched. Auto-approve if the tool is read-only.
-	if tool.Annotations.ReadOnlyHint {
-		runTool()
-		return false
-	}
-
-	// Default: ask the user for confirmation
-	return r.askUserForConfirmation(ctx, sess, toolCall, tool, events, a, runTool)
-}
-
-// permissionChecker pairs a checker with a human-readable source label.
-type permissionChecker struct {
-	checker *permissions.Checker
-	source  string
-}
-
-// permissionCheckers returns the ordered list of permission checkers to evaluate.
-func (r *LocalRuntime) permissionCheckers(sess *session.Session) []permissionChecker {
-	var checkers []permissionChecker
-	if sess.Permissions != nil {
-		checkers = append(checkers, permissionChecker{
-			checker: permissions.NewChecker(&latest.PermissionsConfig{
-				Allow: sess.Permissions.Allow,
-				Ask:   sess.Permissions.Ask,
-				Deny:  sess.Permissions.Deny,
-			}),
-			source: "session permissions",
-		})
-	}
-	if tc := r.team.Permissions(); tc != nil {
-		checkers = append(checkers, permissionChecker{
-			checker: tc,
-			source:  "permissions configuration",
-		})
-	}
-	return checkers
-}
-
-// askUserForConfirmation sends a confirmation event and waits for user response.
-// This is only called when --yolo is not active and no permission rule auto-approved the tool.
-func (r *LocalRuntime) askUserForConfirmation(
-	ctx context.Context,
-	sess *session.Session,
-	toolCall tools.ToolCall,
-	tool tools.Tool,
-	events chan Event,
-	a *agent.Agent,
-	runTool func(),
-) (canceled bool) {
-	toolName := toolCall.Function.Name
-	slog.Debug("Tools not approved, waiting for resume", "tool", toolName, "session_id", sess.ID)
-	events <- ToolCallConfirmation(toolCall, tool, a.Name())
-
-	r.executeOnUserInputHooks(ctx, sess.ID, "tool confirmation")
-
-	select {
-	case req := <-r.resumeChan:
-		switch req.Type {
-		case ResumeTypeApprove:
-			slog.Debug("Resume signal received, approving tool", "tool", toolName, "session_id", sess.ID)
-			runTool()
-		case ResumeTypeApproveSession:
-			slog.Debug("Resume signal received, approving session", "tool", toolName, "session_id", sess.ID)
-			sess.ToolsApproved = true
-			runTool()
-		case ResumeTypeApproveTool:
-			// Add the tool to session's allow list for future auto-approval
-			approvedTool := req.ToolName
-			if approvedTool == "" {
-				approvedTool = toolName
-			}
-			if sess.Permissions == nil {
-				sess.Permissions = &session.PermissionsConfig{}
-			}
-			if !slices.Contains(sess.Permissions.Allow, approvedTool) {
-				sess.Permissions.Allow = append(sess.Permissions.Allow, approvedTool)
-			}
-			slog.Debug("Resume signal received, approving tool permanently", "tool", approvedTool, "session_id", sess.ID)
-			runTool()
-		case ResumeTypeReject:
-			slog.Debug("Resume signal received, rejecting tool", "tool", toolName, "session_id", sess.ID, "reason", req.Reason)
-			rejectMsg := "The user rejected the tool call."
-			if strings.TrimSpace(req.Reason) != "" {
-				rejectMsg += " Reason: " + strings.TrimSpace(req.Reason)
-			}
-			r.addToolErrorResponse(ctx, sess, toolCall, tool, events, a, rejectMsg)
-		}
-		return false
-	case <-ctx.Done():
-		slog.Debug("Context cancelled while waiting for resume", "tool", toolName, "session_id", sess.ID)
-		r.addToolErrorResponse(ctx, sess, toolCall, tool, events, a, "The tool call was canceled by the user.")
-		return true
-	}
-}
-
-// executeToolWithHandler is a common helper that handles tool execution, error handling,
-// event emission, and session updates. It reduces duplication between runTool and runAgentTool.
-func (r *LocalRuntime) executeToolWithHandler(
-	ctx context.Context,
-	toolCall tools.ToolCall,
-	tool tools.Tool,
-	events chan Event,
-	sess *session.Session,
-	a *agent.Agent,
-	spanName string,
-	execute func(ctx context.Context) (*tools.ToolCallResult, time.Duration, error),
-) {
-	ctx, span := r.startSpan(ctx, spanName, trace.WithAttributes(
-		attribute.String("tool.name", toolCall.Function.Name),
-		attribute.String("agent", a.Name()),
-		attribute.String("session.id", sess.ID),
-		attribute.String("tool.call_id", toolCall.ID),
-	))
-	defer span.End()
-
-	events <- ToolCall(toolCall, tool, a.Name())
-
-	res, duration, err := execute(ctx)
-
-	telemetry.RecordToolCall(ctx, toolCall.Function.Name, sess.ID, a.Name(), duration, err)
-
-	if err != nil {
-		if errors.Is(err, context.Canceled) || errors.Is(ctx.Err(), context.Canceled) {
-			slog.Debug("Tool handler canceled by context", "tool", toolCall.Function.Name, "agent", a.Name(), "session_id", sess.ID)
-			res = tools.ResultError("The tool call was canceled by the user.")
-			span.SetStatus(codes.Ok, "tool handler canceled by user")
-		} else {
-			span.RecordError(err)
-			span.SetStatus(codes.Error, "tool handler error")
-			slog.Error("Error calling tool", "tool", toolCall.Function.Name, "error", err)
-			res = tools.ResultError(fmt.Sprintf("Error calling tool: %v", err))
-		}
-	} else {
-		span.SetStatus(codes.Ok, "tool handler completed")
-		slog.Debug("Tool call completed", "tool", toolCall.Function.Name, "output_length", len(res.Output))
-	}
-
-	events <- ToolCallResponse(toolCall, tool, res, res.Output, a.Name())
-
-	// Ensure tool response content is not empty for API compatibility
-	content := res.Output
-	if strings.TrimSpace(content) == "" {
-		content = "(no output)"
-	}
-
-	toolResponseMsg := chat.Message{
-		Role:       chat.MessageRoleTool,
-		Content:    content,
-		ToolCallID: toolCall.ID,
-		IsError:    res.IsError,
-		CreatedAt:  time.Now().Format(time.RFC3339),
-	}
-
-	// If the tool result contains images, attach them as MultiContent
-	if len(res.Images) > 0 {
-		multiContent := []chat.MessagePart{
-			{
-				Type: chat.MessagePartTypeText,
-				Text: content,
-			},
-		}
-		for _, img := range res.Images {
-			multiContent = append(multiContent, chat.MessagePart{
-				Type: chat.MessagePartTypeImageURL,
-				ImageURL: &chat.MessageImageURL{
-					URL:    "data:" + img.MimeType + ";base64," + img.Data,
-					Detail: chat.ImageURLDetailAuto,
-				},
-			})
-		}
-		toolResponseMsg.MultiContent = multiContent
-	}
-
-	addAgentMessage(sess, a, &toolResponseMsg, events)
-}
-
-// runTool executes agent tools from toolsets (MCP, filesystem, etc.).
-func (r *LocalRuntime) runTool(ctx context.Context, tool tools.Tool, toolCall tools.ToolCall, events chan Event, sess *session.Session, a *agent.Agent) {
-	// Get hooks executor for this agent
-	hooksExec := r.getHooksExecutor(a)
-
-	// Execute pre-tool hooks if configured
-	if hooksExec != nil && hooksExec.HasPreToolUseHooks() {
-		toolInput := parseToolInput(toolCall.Function.Arguments)
-		input := &hooks.Input{
-			SessionID: sess.ID,
-			Cwd:       r.workingDir,
-			ToolName:  toolCall.Function.Name,
-			ToolUseID: toolCall.ID,
-			ToolInput: toolInput,
-		}
-
-		result, err := hooksExec.ExecutePreToolUse(ctx, input)
-		switch {
-		case err != nil:
-			slog.Warn("Pre-tool hook execution failed", "tool", toolCall.Function.Name, "error", err)
-		case !result.Allowed:
-			// Hook blocked the tool call
-			slog.Debug("Pre-tool hook blocked tool call", "tool", toolCall.Function.Name, "message", result.Message)
-			events <- HookBlocked(toolCall, tool, result.Message, a.Name())
-			r.addToolErrorResponse(ctx, sess, toolCall, tool, events, a, "Tool call blocked by hook: "+result.Message)
-			return
-		default:
-			if result.SystemMessage != "" {
-				events <- Warning(result.SystemMessage, a.Name())
-			}
-			if result.ModifiedInput != nil {
-				updated, merr := json.Marshal(result.ModifiedInput)
-				if merr != nil {
-					slog.Warn("Failed to marshal modified tool input from hook", "tool", toolCall.Function.Name, "error", merr)
-				} else {
-					slog.Debug("Pre-tool hook modified tool input", "tool", toolCall.Function.Name)
-					toolCall.Function.Arguments = string(updated)
-				}
-			}
-		}
-	}
-
-	r.executeToolWithHandler(ctx, toolCall, tool, events, sess, a, "runtime.tool.handler",
-		func(ctx context.Context) (*tools.ToolCallResult, time.Duration, error) {
-			res, err := tool.Handler(ctx, toolCall)
-			return res, 0, err
-		})
-
-	// Execute post-tool hooks if configured
-	if hooksExec != nil && hooksExec.HasPostToolUseHooks() {
-		toolInput := parseToolInput(toolCall.Function.Arguments)
-		input := &hooks.Input{
-			SessionID:    sess.ID,
-			Cwd:          r.workingDir,
-			ToolName:     toolCall.Function.Name,
-			ToolUseID:    toolCall.ID,
-			ToolInput:    toolInput,
-			ToolResponse: nil, // TODO: pass actual tool response if needed
-		}
-
-		result, err := hooksExec.ExecutePostToolUse(ctx, input)
-		if err != nil {
-			slog.Warn("Post-tool hook execution failed", "tool", toolCall.Function.Name, "error", err)
-		} else if result.SystemMessage != "" {
-			events <- Warning(result.SystemMessage, a.Name())
-		}
-	}
-}
-
-// parseToolInput parses tool arguments JSON into a map
-func parseToolInput(arguments string) map[string]any {
-	var result map[string]any
-	if err := json.Unmarshal([]byte(arguments), &result); err != nil {
-		return nil
-	}
-	return result
-}
-
-func (r *LocalRuntime) runAgentTool(ctx context.Context, handler ToolHandlerFunc, sess *session.Session, toolCall tools.ToolCall, tool tools.Tool, events chan Event, a *agent.Agent) {
-	r.executeToolWithHandler(ctx, toolCall, tool, events, sess, a, "runtime.tool.handler.runtime",
-		func(ctx context.Context) (*tools.ToolCallResult, time.Duration, error) {
-			start := time.Now()
-			res, err := handler(ctx, sess, toolCall, events)
-			return res, time.Since(start), err
-		})
-}
-
-func addAgentMessage(sess *session.Session, a *agent.Agent, msg *chat.Message, events chan Event) {
-	agentMsg := session.NewAgentMessage(a, msg)
-	sess.AddMessage(agentMsg)
-	events <- MessageAdded(sess.ID, agentMsg, a.Name())
-}
-
-// addToolErrorResponse adds a tool error response to the session and emits the event.
-// This consolidates the common pattern used by validation, rejection, and cancellation responses.
-func (r *LocalRuntime) addToolErrorResponse(_ context.Context, sess *session.Session, toolCall tools.ToolCall, tool tools.Tool, events chan Event, a *agent.Agent, errorMsg string) {
-	events <- ToolCallResponse(toolCall, tool, tools.ResultError(errorMsg), errorMsg, a.Name())
-
-	toolResponseMsg := chat.Message{
-		Role:       chat.MessageRoleTool,
-		Content:    errorMsg,
-		ToolCallID: toolCall.ID,
-		IsError:    true,
-		CreatedAt:  time.Now().Format(time.RFC3339),
-	}
-	addAgentMessage(sess, a, &toolResponseMsg, events)
-}
-
-// startSpan wraps tracer.Start, returning a no-op span if the tracer is nil.
 func (r *LocalRuntime) startSpan(ctx context.Context, name string, opts ...trace.SpanStartOption) (context.Context, trace.Span) {
 	if r.tracer == nil {
 		return ctx, trace.SpanFromContext(ctx)
@@ -2148,82 +1041,4 @@ func (r *LocalRuntime) elicitationHandler(ctx context.Context, req *mcp.ElicitPa
 		slog.Debug("Context cancelled while waiting for elicitation response")
 		return tools.ElicitationResult{}, ctx.Err()
 	}
-}
-
-// stripImageContent returns a copy of messages with all image-related content
-// removed. This is used when the target model doesn't support image input to
-// prevent API errors. Text content is preserved; image parts in MultiContent
-// are filtered out, and file attachments with image MIME types are dropped.
-func stripImageContent(messages []chat.Message) []chat.Message {
-	result := make([]chat.Message, len(messages))
-	for i, msg := range messages {
-		result[i] = msg
-
-		if len(msg.MultiContent) == 0 {
-			continue
-		}
-
-		var filtered []chat.MessagePart
-		for _, part := range msg.MultiContent {
-			switch part.Type {
-			case chat.MessagePartTypeImageURL:
-				// Drop image URL parts entirely.
-				continue
-			case chat.MessagePartTypeFile:
-				// Drop file parts that are images.
-				if part.File != nil && chat.IsImageMimeType(part.File.MimeType) {
-					continue
-				}
-			}
-			filtered = append(filtered, part)
-		}
-
-		if len(filtered) != len(msg.MultiContent) {
-			result[i].MultiContent = filtered
-			slog.Debug("Stripped image content from message", "role", msg.Role, "original_parts", len(msg.MultiContent), "remaining_parts", len(filtered))
-		}
-	}
-	return result
-}
-
-// charsPerToken is the average number of characters per token used for
-// estimation. A value of 4 is a widely-used heuristic for English text;
-// it slightly overestimates token counts for code/JSON (which is ~3.5),
-// making compaction trigger earlier — the safe direction.
-const charsPerToken = 4
-
-// estimateMessageTokens returns a rough token-count estimate for a single
-// chat message based on its text length. This is intentionally conservative
-// (overestimates) so that proactive compaction fires before we hit the limit.
-// The estimate includes the message content, multi-content text parts, and
-// a small overhead per message for role/metadata tokens.
-func estimateMessageTokens(msg *chat.Message) int64 {
-	var chars int
-
-	// Primary text content.
-	chars += len(msg.Content)
-
-	// Multi-content parts (e.g., tool results with image descriptions).
-	for _, part := range msg.MultiContent {
-		chars += len(part.Text)
-	}
-
-	// Reasoning / thinking content.
-	chars += len(msg.ReasoningContent)
-
-	// Tool call arguments (they count toward input tokens on the next turn).
-	for _, tc := range msg.ToolCalls {
-		chars += len(tc.Function.Arguments)
-		chars += len(tc.Function.Name)
-	}
-
-	// Per-message overhead: role, ToolCallID, delimiters, etc.
-	// Models typically use 3-7 tokens for message framing.
-	const perMessageOverhead = 5
-
-	if chars == 0 {
-		return perMessageOverhead
-	}
-
-	return int64(chars/charsPerToken) + perMessageOverhead
 }
